alert_service/requirements.txt
python-telegram-bot==20.3
redis==5.2.0
